resource: support cloudformation stack type
This allows you to define a cloudformation stack like:
```
- Type: "Cloudformation"
  Path: "examples/cloudformation/dynamo/table.yml"
  Name: "painter3"
  CloudformationParameters:
    - "HashKeyElementName=Painter"
```

We use the Name and path to ensure that stacks are unique.

Cloudformation support depends on change sets that we use to update a
deployed stack as the cloudformation changes.

The Path for cloudformation should go directly to the template you want
to use. Unlike terraform and CDK where there is a directory of contents.