resource account: allow setting NoStackInheritance
You can now set NoStackInheritance: true for an account. This will make
it so that a stack doesn't inherit from the current OU or parent OUs.

The most useful case for this is when you have the management account at
the top level and want to declare top level stacks as well to target all
sub accounts except the management account.